Reorganize demos: notebooks/, examples/, scripts/ split
- Move chirp.py into the fuge package (src/fuge/chirp.py) so it's
  importable from anywhere without sys.path hacks.
- Split examples/ by purpose: exploratory demos to notebooks/ (destined
  to become .ipynb), ML pipeline reference scripts stay in examples/,
  empty scripts/ added for future optimization tools.
- Update imports from `chirp` to `fuge.chirp`; drop now-redundant
  sys.path.insert(0, ".") lines.
