+++ notebooks/spectral_demo.py
